Fixed comments

+++ deploy/cleanup.sh
@@ -64,15 +64,18 @@ if [[ ! -z "$PRE_PROVISIONED_BATCH_ACCOUNT_NAME" ]] && [ "${NO_BATCH_ACCOUNT_POO
     BATCH_ACCOUNT_ID=$(az batch account list --query "[?name == '${PRE_PROVISIONED_BATCH_ACCOUNT_NAME}'].id" -o tsv)
     BATCH_ACCOUNT_RG_NAME=$(az resource show --ids ${BATCH_ACCOUNT_ID} --query resourceGroup -o tsv)
     BATCH_ACCOUNT_KEY=$(az batch account keys list --name ${PRE_PROVISIONED_BATCH_ACCOUNT_NAME} --resource-group ${BATCH_ACCOUNT_RG_NAME} | jq ".primary")
-    echo "Cleanup the batch pool: ${BATCH_ACCOUNT_POOL_NAME}"
-    az batch pool delete --yes \
-        --pool-id ${BATCH_ACCOUNT_POOL_NAME} \
-        --account-name ${PRE_PROVISIONED_BATCH_ACCOUNT_NAME} \
-        --account-key ${BATCH_ACCOUNT_KEY}
 
     echo "Cleaning the batch job: ${BATCH_ACCOUNT_POOL_NAME}"
     az batch job delete --yes \
         --job-id ${BATCH_ACCOUNT_POOL_NAME} \
         --account-name ${PRE_PROVISIONED_BATCH_ACCOUNT_NAME} \
         --account-key ${BATCH_ACCOUNT_KEY}
+    if [[ $? ==0 ]]; then
+        echo "Cleanup the batch pool: ${BATCH_ACCOUNT_POOL_NAME}"
+        az batch pool delete --yes \
+            --pool-id ${BATCH_ACCOUNT_POOL_NAME} \
+            --account-name ${PRE_PROVISIONED_BATCH_ACCOUNT_NAME} \
+            --account-key ${BATCH_ACCOUNT_KEY}
+    fi
+
 fi
